refactor: SpringRestDocs 설정 로직 분리, 어노테이션 추상화
