+++ src/test/java/donggi/dev/kkeuroolryo/IntegrationTest.java
@@ -1,0 +1,16 @@
+package donggi.dev.kkeuroolryo;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.junit.jupiter.api.Tag;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Tag("integration")
+public @interface IntegrationTest {
+
+}
