+++ src/test/java/donggi/dev/kkeuroolryo/web/question/QuestionRestControllerRestDocsTest.java
@@ -1,75 +1,23 @@
 package donggi.dev.kkeuroolryo.web.question;
 
 import static io.restassured.RestAssured.given;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.restdocs.restassured.RestAssuredRestDocumentation.document;
-import static org.springframework.restdocs.restassured.RestAssuredRestDocumentation.documentationConfiguration;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import donggi.dev.kkeuroolryo.InitRestDocsTest;
+import donggi.dev.kkeuroolryo.RestAssuredAndRestDocsTest;
 import donggi.dev.kkeuroolryo.web.question.dto.QuestionRegisterRequest;
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.config.ObjectMapperConfig;
-import io.restassured.config.RestAssuredConfig;
-import io.restassured.internal.mapping.Jackson2Mapper;
-import io.restassured.specification.RequestSpecification;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.restdocs.payload.JsonFieldType;
-import org.springframework.test.context.ActiveProfiles;
 
 @DisplayName("API 문서화 : 유저 골라바 질문 등록 요청")
-@ActiveProfiles("test")
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@ExtendWith(RestDocumentationExtension.class)
-class QuestionRestControllerRestDocsTest {
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @LocalServerPort
-    private int port;
-
-    private RequestSpecification spec;
-
-    {
-        setUpRestAssured();
-    }
-
-    @BeforeEach
-    void setUpRestDocs(RestDocumentationContextProvider restDocumentation) {
-        RestAssured.port = port;
-
-        this.spec = new RequestSpecBuilder()
-            .setPort(port)
-            .addFilter(documentationConfiguration(restDocumentation)
-                .operationPreprocessors()
-                .withRequestDefaults(prettyPrint())
-                .withResponseDefaults(prettyPrint()))
-            .build();
-    }
-
-    private void setUpRestAssured() {
-        RestAssured.config = RestAssuredConfig.config()
-            .objectMapperConfig(
-                new ObjectMapperConfig(
-                    new Jackson2Mapper((type, charset) -> objectMapper)
-                )
-            );
-    }
+@RestAssuredAndRestDocsTest
+class QuestionRestControllerRestDocsTest extends InitRestDocsTest {
 
     @Test
     @DisplayName("골라바 질문 등록 요청이 정상적인 경우 질문 생성 후 상태코드를 반환한다.")
